implementation of negamax

+++ src/bot/evaluation/mod.rs
@@ -8,8 +8,8 @@ use crate::game::{
 
 use super::Board;
 
-const SCORE_CASTLE_RIGHT: i32 = 40;
-const CASTLE_SCORE_BY_INDEX: [i32; 16] = [
+const SCORE_CASTLE_RIGHT: i64 = 40;
+const CASTLE_SCORE_BY_INDEX: [i64; 16] = [
     0,                      //0000
     SCORE_CASTLE_RIGHT,     //0001
     SCORE_CASTLE_RIGHT,     //0010
@@ -28,13 +28,13 @@ const CASTLE_SCORE_BY_INDEX: [i32; 16] = [
     0,                      //1111
 ];
 
-const SCORE_PAWN: i32 = 10;
-const SCORE_KNIGHT: i32 = 30;
-const SCORE_BISHOP: i32 = 30;
-const SCORE_ROOK: i32 = 50;
-const SCORE_QUEEN: i32 = 90;
+const SCORE_PAWN: i64 = 10;
+const SCORE_KNIGHT: i64 = 30;
+const SCORE_BISHOP: i64 = 30;
+const SCORE_ROOK: i64 = 50;
+const SCORE_QUEEN: i64 = 90;
 
-const SCORE_PIECES: [i32; 6] = [
+const SCORE_PIECES: [i64; 6] = [
     SCORE_PAWN,
     SCORE_KNIGHT,
     SCORE_BISHOP,
@@ -43,9 +43,10 @@ const SCORE_PIECES: [i32; 6] = [
     1, //multiplier for king (1 king * 1 = 1) Score won't be effected by this
 ];
 
-const MOBILITY_SCORE: i32 = 2;
+const MOBILITY_SCORE: i64 = 2;
 
-pub fn evaluate_board(board: &Board) -> i32 {
+/// Evaluates the board state and returns a score. Positive score indicates white is winning, negative score indicates black is winning.
+pub fn evaluate_board(board: &Board) -> i64 {
     let mut score = 0;
 
     let board_state = board.cur_state();
@@ -55,19 +56,19 @@ pub fn evaluate_board(board: &Board) -> i32 {
     score += evaluate_castle(&board_state.castling_rights);
 
     //mobility score
-    let move_count = MoveGeneration::generate_legal_moves(board).len() as i32;
-    score += move_count * MOBILITY_SCORE * color.perspective() as i32;
+    let move_count = MoveGeneration::generate_legal_moves(board).len() as i64;
+    score += move_count * MOBILITY_SCORE * color.perspective() as i64;
 
     score / 10
 }
 
 #[inline(always)]
-fn evaluate_castle(castle_rights: &CastleRights) -> i32 {
+fn evaluate_castle(castle_rights: &CastleRights) -> i64 {
     CASTLE_SCORE_BY_INDEX[castle_rights.as_u8() as usize]
 }
 
 #[inline(always)]
-fn evaluate_pieces(board: &Board) -> i32 {
+fn evaluate_pieces(board: &Board) -> i64 {
     let mut score = 0;
 
     let bb = board.get_bb_pieces();
@@ -76,15 +77,15 @@ fn evaluate_pieces(board: &Board) -> i32 {
         let white_piece_board = bb[Color::White][piece_type];
         let black_piece_board = bb[Color::Black][piece_type];
         score += piece_score
-            * (white_piece_board.count_ones() as i32 - black_piece_board.count_ones() as i32);
+            * (white_piece_board.count_ones() as i64 - black_piece_board.count_ones() as i64);
 
         iter_set_bits(*white_piece_board).for_each(|sq| {
             let index = Color::White.relative_sq(sq);
-            score += piece_square_table::PIECE_SQUARE_TABLES[piece_type][index as usize] as i32;
+            score += piece_square_table::PIECE_SQUARE_TABLES[piece_type][index as usize] as i64;
         });
         iter_set_bits(*black_piece_board).for_each(|sq| {
             let index = Color::Black.relative_sq(sq);
-            score -= piece_square_table::PIECE_SQUARE_TABLES[piece_type][index as usize] as i32;
+            score -= piece_square_table::PIECE_SQUARE_TABLES[piece_type][index as usize] as i64;
         });
     }
     score
